libroach: further decomposition of db.cc
Split the various engine implementations out of `db.cc`. Move the merge
operator routines into `merge.cc`. Move the `MVCC*` routines into
`mvcc.cc`. Move the custom RocksDB comparator to `comparator.cc`. Move
the cache routines into `cache.cc`. Move the options routines into
`options.cc`.

This is all code movement: no functional changes.

Release note: None
