+++ c-deps/libroach/getter.h
@@ -1,0 +1,59 @@
+// Copyright 2017 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied.  See the License for the specific language governing
+// permissions and limitations under the License.
+
+#pragma once
+
+#include <libroach.h>
+#include <rocksdb/db.h>
+#include <rocksdb/iterator.h>
+
+namespace cockroach {
+
+// Getter defines an interface for retrieving a value from either an
+// iterator or an engine. It is used by ProcessDeltaKey to abstract
+// whether the "base" layer is an iterator or an engine.
+struct Getter {
+  virtual DBStatus Get(DBString* value) = 0;
+};
+
+// IteratorGetter is an implementation of the Getter interface which
+// retrieves the value currently pointed to by the supplied
+// iterator. It is ok for the supplied iterator to be NULL in which
+// case no value will be retrieved.
+struct IteratorGetter : public Getter {
+  rocksdb::Iterator* const base;
+
+  IteratorGetter(rocksdb::Iterator* iter) : base(iter) {}
+
+  virtual DBStatus Get(DBString* value);
+};
+
+// DBGetter is an implementation of the Getter interface which
+// retrieves the value for the supplied key from a rocksdb::DB.
+struct DBGetter : public Getter {
+  rocksdb::DB* const rep;
+  rocksdb::ReadOptions const options;
+  std::string const key;
+
+  DBGetter(rocksdb::DB* const r, rocksdb::ReadOptions opts, std::string&& k)
+      : rep(r), options(opts), key(std::move(k)) {}
+
+  virtual DBStatus Get(DBString* value);
+};
+
+}  // namespace cockroach
+
+// local variables:
+// mode: c++
+// end:
